refactor: 인증 확인 훅 useUser로 이름 변경
